requirements.txt
absl-py==2.1.0
affine==2.4.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.115.6
filelock==3.16.1
fiona==1.9.6
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.12.0
gast==0.6.0
geopandas==0.14.4
google-pasta==0.2.0
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.60.0
numpy==1.24.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.59.6
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
rasterio==1.3.10
regex==2024.11.6
requests==2.32.3
rich==13.9.4
Rtree==1.2.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
soundfile==0.13.0
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.5.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.2
wrapt==1.16.0
